Type mitigation

+++ server/src/main/java/fi/chaoscompany/server/models/GameObject.java
@@ -1,30 +1,30 @@
 package fi.chaoscompany.server.models;
 
 public class GameObject {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
     public GameObject() {
     }
 
-    public GameObject(int x, int y) {
+    public GameObject(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 }
